# schema/openapi.yaml
openapi: 3.0.3
info:
  title: Agenty API
  version: v1
  description: |
    The API of an agenty server: it stores harnesses, runs them, and streams
    each run's events, approval requests included.

    Every request signs in with a user's bearer token. Harnesses and runs
    live in workspaces; a workspace the user is not a member of is not found,
    and neither is a run asked for in another workspace than its own. The
    exception: a user still reads the runs of conversations they started in
    a workspace they left, read-only.

    Errors are returned as an Error: 400 for an invalid request, 401 without
    a valid token, 403 for a request for a host that is not local, from a
    web page of an origin that is not configured, or for the audit log's
    runs, events and export by a user who is not an auditor, 404 for anything not
    found, workspaces included, 409 for cancelling a run that has finished,
    for following up a run that has not or that is followed up already, or
    for a run that has not finished but that this server does not run, 415
    for a body that is not JSON, 500 when the server fails, and 503 while
    the server is stopping.
    A run that cannot start, such as one whose grants no MCP server serves,
    is queued all the same and fails.

    This file is the API's source of truth: the Go types in internal/api and
    the server interface in internal/server are generated from it.
servers:
  - url: http://127.0.0.1:8080
security:
  - bearerAuth: []
paths:
...
  /v1/conversations:
    get:
      operationId: listConversations
      summary: The conversations the user started, latest activity first, a page at a time
      description: |
        The conversations whose first run the user started, in every
        workspace, those they left included, the one whose latest run was
        started last first. Only its starter continues, cancels and answers
        a conversation, through its workspace. A conversation followed up
        while the pages are read moves to the first page, so later pages
        leave it out.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - name: before
          in: query
          description: Continue after the conversation this cursor names, as ConversationList.next gives it.
          schema: {type: string}
      responses:
        "200":
          description: A page of conversations.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ConversationList"}
        default: {$ref: "#/components/responses/Error"}
  /v1/conversations/{id}:
    get:
      operationId: getConversation
      summary: A conversation the user started, with its runs
      description: |
        Only the user who started the conversation finds it, in any
        workspace, those they left or the config no longer has included;
        anyone else, other members and auditors too, finds no such
        conversation. One of a workspace the user left is read-only. Only
        the ID of its first run names it. Its runs come oldest first; its
        latest is followed, continued, cancelled and answered through its
        workspace.
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the conversation's first run.
          schema: {type: string}
      responses:
        "200":
          description: The conversation and its runs.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Conversation"}
        default: {$ref: "#/components/responses/Error"}
...
  /v1/workspaces/{workspace}:
    parameters:
      - $ref: "#/components/parameters/Workspace"
    get:
      operationId: getWorkspace
      summary: The workspace and its members
      responses:
        "200":
          description: The workspace.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/WorkspaceDetail"}
        default: {$ref: "#/components/responses/Error"}
...
  /v1/workspaces/{workspace}/harnesses:
    parameters:
      - $ref: "#/components/parameters/Workspace"
    get:
      operationId: listHarnesses
      summary: The latest version of every harness, sorted by name
      responses:
        "200":
          description: The harnesses.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/HarnessVersion"}
        default: {$ref: "#/components/responses/Error"}
  /v1/workspaces/{workspace}/harnesses/{name}:
    parameters:
      - $ref: "#/components/parameters/Workspace"
      - name: name
        in: path
        required: true
        schema: {type: string}
    get:
      operationId: getHarness
      summary: The latest version of a harness
      responses:
        "200":
          description: The harness version.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/HarnessVersion"}
        default: {$ref: "#/components/responses/Error"}
    put:
      operationId: putHarness
      summary: Store a harness as its next version
      description: |
        A changed harness becomes the next version; an unchanged one is left
        as it is, and its latest version is returned. The name must match
        the path.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/Harness"}
      responses:
        "200":
          description: The stored version.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/HarnessVersion"}
        default: {$ref: "#/components/responses/Error"}
  /v1/workspaces/{workspace}/runs:
    parameters:
      - $ref: "#/components/parameters/Workspace"
    post:
      operationId: createRun
      summary: Start a run of a harness's latest version
      description: |
        Queues the run, which starts once a worker is free; its events tell
        when. A run whose MCP servers cannot serve its grants fails.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateRun"}
      responses:
        "201":
          description: The run, queued.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Run"}
        default: {$ref: "#/components/responses/Error"}
...
  /v1/workspaces/{workspace}/runs/{id}/follow-up:
    parameters:
      - $ref: "#/components/parameters/Workspace"
      - $ref: "#/components/parameters/RunID"
    post:
      operationId: followUpRun
      summary: Continue a run's conversation with a new run
      description: |
        Starts a run of the harness version of the run it follows, not the
        latest, that continues the conversation: the model sees the earlier
        runs' transcripts, with secrets redacted, before the new input. Only the conversation's
        latest run can be followed up, once it has finished; any other is a
        conflict. A run that failed or was cancelled is continued from where
        it stopped: the model is told so, and that a tool call whose result
        was not recorded may or may not have run. No call is run again.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/FollowUp"}
      responses:
        "201":
          description: The run, queued.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Run"}
        default: {$ref: "#/components/responses/Error"}
  /v1/workspaces/{workspace}/runs/{id}/cancel:
    parameters:
      - $ref: "#/components/parameters/Workspace"
      - $ref: "#/components/parameters/RunID"
    post:
      operationId: cancelRun
      summary: Cancel a queued, waiting or running run
      description: |
        A queued run, or one waiting for approval, ends at once; a running
        one as soon as what it is doing stops. Either ends as cancelled by
        the user; its events tell when. A run that finishes as it is
        cancelled may keep its own end.
      responses:
        "202":
          description: The run is being cancelled.
        default: {$ref: "#/components/responses/Error"}
  /v1/workspaces/{workspace}/runs/{id}/transcript:
    parameters:
      - $ref: "#/components/parameters/Workspace"
      - $ref: "#/components/parameters/RunID"
    get:
      operationId: getRunTranscript
      summary: A run's conversation with the model, with secrets redacted
      responses:
        "200":
          description: The messages, in order.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/TranscriptMessage"}
        default: {$ref: "#/components/responses/Error"}
...
  /v1/workspaces/{workspace}/runs/{id}/approvals/{approval}:
    parameters:
      - $ref: "#/components/parameters/Workspace"
      - $ref: "#/components/parameters/RunID"
      - name: approval
        in: path
        required: true
        schema: {type: string}
    post:
      operationId: answerApproval
      summary: Answer an approval request
      description: |
        The signed-in user is recorded as the approver. The answer queues the
        run, which resumes at the call; a request is answered once, and not
        after it expired, which rejected it.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/Answer"}
      responses:
        "204":
          description: The answer is accepted.
        default: {$ref: "#/components/responses/Error"}
  /v1/workspaces/{workspace}/approvals:
    parameters:
      - $ref: "#/components/parameters/Workspace"
    get:
      operationId: listApprovals
      summary: The approval requests the user's own runs wait for, oldest first
      responses:
        "200":
          description: The requests.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/ApprovalRequest"}
        default: {$ref: "#/components/responses/Error"}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Workspace:
      name: workspace
      in: path
      required: true
      schema: {type: string}
    RunID:
      name: id
      in: path
      required: true
      description: |
        A run of a conversation the user started. Any other run, also one of
        another member's conversation, is not found.
      schema: {type: string}
...
    Limit:
      name: limit
      in: query
      description: The page size.
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
  responses:
    Error:
      description: An error.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: {type: string}
...
    WorkspaceDetail:
      description: A workspace, as its members see it.
      type: object
      additionalProperties: false
      required: [name, members]
      properties:
        name: {type: string}
        members:
          description: The users who are members of it, sorted.
          type: array
          items: {type: string}
    Harness:
      description: A declarative definition of an agent.
      type: object
      additionalProperties: false
      required: [name, instructions, model, limits]
      properties:
        name: {type: string}
        instructions: {type: string}
        model: {$ref: "#/components/schemas/HarnessModel"}
        tools:
          description: The granted tools.
          type: array
          items: {type: string}
        limits: {$ref: "#/components/schemas/Limits"}
        policy:
          description: |
            Rego modules in package agenty.tool that tighten central policy
            for this harness.
          type: array
          items: {$ref: "#/components/schemas/PolicyModule"}
    HarnessModel:
      type: object
      additionalProperties: false
      required: [provider, name]
      properties:
        provider: {type: string}
        name: {type: string}
    Limits:
      type: object
      additionalProperties: false
      required: [max_steps, max_tool_calls]
      properties:
        max_steps:
          description: Bounds the model calls of a run.
          type: integer
        max_tool_calls:
          description: Bounds the tool calls of a run, denied ones included.
          type: integer
    PolicyModule:
      type: object
      additionalProperties: false
      required: [name, source]
      properties:
        name: {type: string}
        source: {type: string}
    HarnessVersion:
      type: object
      required: [version, harness, created_at]
      properties:
        version: {type: integer}
        harness: {$ref: "#/components/schemas/Harness"}
        created_at: {type: string, format: date-time}
    CreateRun:
      type: object
      additionalProperties: false
      required: [harness, input]
      properties:
        harness:
          description: The harness to run, by name; the run uses its latest version.
          type: string
        input: {type: string}
    FollowUp:
      type: object
      additionalProperties: false
      required: [input]
      properties:
        input: {type: string}
    RunStatus:
      type: string
      enum: [queued, running, waiting, succeeded, failed, cancelled]
    Run:
      type: object
      required: [id, conversation_id, harness, harness_version, started_by, input, status, output, steps, usage, created_at]
      properties:
        id: {type: string}
        conversation_id:
          description: |
            The conversation the run belongs to, named by the ID of its
            first run, which may be this one.
          type: string
        follows:
          description: The run whose conversation this run continues, if any.
          type: string
        harness: {type: string}
        harness_version: {type: integer}
        started_by:
          description: The user who started the run.
          type: string
        input: {type: string}
        status: {$ref: "#/components/schemas/RunStatus"}
        output: {type: string}
        steps: {type: integer}
        usage:
          description: |
            The tokens of the model calls that wrote the run's replies so
            far. A call that failed, such as one whose reply was cut off, is
            not counted.
          allOf: [{$ref: "#/components/schemas/Usage"}]
        error:
          description: Why the run failed, or who cancelled it.
          type: string
        created_at: {type: string, format: date-time}
        finished_at:
          description: Unset until the run has finished.
          type: string
          format: date-time
          x-omitzero: true
    Usage:
      description: |
        The tokens of model calls. Input the provider read from its prompt
        cache, or wrote to it, is counted apart from the rest of the input.
      type: object
      required: [input_tokens, output_tokens, cache_write_tokens, cache_read_tokens]
      properties:
        input_tokens:
          description: Input tokens neither read from nor written to the cache.
          type: integer
          format: int64
        output_tokens: {type: integer, format: int64}
        cache_write_tokens:
          description: Input tokens written to the prompt cache.
          type: integer
          format: int64
        cache_read_tokens:
          description: Input tokens read from the prompt cache.
          type: integer
          format: int64
...
    ConversationSummary:
      type: object
      required: [id, workspace, harness, title, status]
      properties:
        id:
          description: The ID of the conversation's first run, which names it.
          type: string
        workspace: {type: string}
        harness: {type: string}
        title:
          description: The first run's input, cut to its first 100 characters.
          type: string
        status:
          description: The status of the conversation's latest run.
          allOf: [{$ref: "#/components/schemas/RunStatus"}]
    Conversation:
      allOf:
        - {$ref: "#/components/schemas/ConversationSummary"}
        - type: object
          required: [runs]
          properties:
            runs:
              description: The conversation's runs, oldest first.
              type: array
              items: {$ref: "#/components/schemas/Run"}
    ConversationList:
      type: object
      required: [conversations]
      properties:
        conversations:
          type: array
          items: {$ref: "#/components/schemas/ConversationSummary"}
        next:
          description: The before parameter of the next page, which may be empty.
          type: string
...
    Role:
      type: string
      enum: [user, assistant]
    ToolCall:
      type: object
      required: [id, name]
      properties:
        id: {type: string}
        name: {type: string}
        args: {$ref: "#/components/schemas/JSON"}
    ToolResult:
      type: object
      required: [call_id, content]
      properties:
        call_id: {type: string}
        content: {type: string}
        is_error: {type: boolean}
    TranscriptMessage:
      description: |
        One message of a run's conversation with the model: the input, a
        model reply, or the results of the reply's tool calls. What the
        provider kept of a reply for the model, such as its thinking, is
        not part of it.
      type: object
      additionalProperties: false
      required: [position, role]
      properties:
        position:
          description: The message's index in the conversation.
          type: integer
        role: {$ref: "#/components/schemas/Role"}
        text: {type: string}
        tool_calls:
          type: array
          items: {$ref: "#/components/schemas/ToolCall"}
        tool_results:
          type: array
          items: {$ref: "#/components/schemas/ToolResult"}
    ApprovalRequest:
      description: |
        A call that policy marked as requiring approval. Secrets are already
        redacted from it.
      type: object
      required: [id, run_id, harness, tool, args, reasons, created_at, expires_at]
      properties:
        id:
          description: Identifies the request when answering it.
          type: string
        run_id: {type: string}
        harness: {type: string}
        tool: {type: string}
        args: {$ref: "#/components/schemas/JSON"}
        reasons:
          type: array
          items: {type: string}
        created_at: {type: string, format: date-time}
        expires_at:
          description: When the request is rejected if no one answers it.
          type: string
          format: date-time
    Answer:
      type: object
      additionalProperties: false
      required: [approved]
      properties:
        approved: {type: boolean}
        reason: {type: string}
    JSON:
      description: Any JSON value, kept as written.
      x-go-type: json.RawMessage